use Long insead of Integer,fix test bug

+++ core-test/src/main/java/org/openstack4j/api/image/v2/ImageV2Tests.java
@@ -1,16 +1,8 @@
 package org.openstack4j.api.image.v2;
 
-import org.openstack4j.api.AbstractTest;
-import org.openstack4j.api.Builders;
-import org.openstack4j.model.common.ActionResponse;
-import org.openstack4j.model.common.Payload;
-import org.openstack4j.model.common.Payloads;
-import org.openstack4j.model.image.v2.ContainerFormat;
-import org.openstack4j.model.image.v2.DiskFormat;
-import org.openstack4j.model.image.v2.Image;
-import org.openstack4j.model.image.v2.Member;
-import org.openstack4j.model.image.v2.Task;
-import org.testng.annotations.Test;
+import static org.testng.Assert.assertEquals;
+import static org.testng.Assert.assertNotNull;
+import static org.testng.Assert.assertTrue;
 
 import java.io.ByteArrayInputStream;
 import java.io.File;
@@ -23,9 +15,17 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import static org.testng.Assert.assertEquals;
-import static org.testng.Assert.assertNotNull;
-import static org.testng.Assert.assertTrue;
+import org.openstack4j.api.AbstractTest;
+import org.openstack4j.api.Builders;
+import org.openstack4j.model.common.ActionResponse;
+import org.openstack4j.model.common.Payload;
+import org.openstack4j.model.common.Payloads;
+import org.openstack4j.model.image.v2.ContainerFormat;
+import org.openstack4j.model.image.v2.DiskFormat;
+import org.openstack4j.model.image.v2.Image;
+import org.openstack4j.model.image.v2.Member;
+import org.openstack4j.model.image.v2.Task;
+import org.testng.annotations.Test;
 
 /**
  * @author emjburns
@@ -77,8 +77,8 @@ public class ImageV2Tests extends AbstractTest {
         String name = "amphora-x64-haproxy";
         ContainerFormat cf = ContainerFormat.BARE;
         DiskFormat df = DiskFormat.QCOW2;
-        Integer mindisk = 0;
-        Integer minram = 0;
+        Long mindisk = 0L;
+        Long minram = 0L;
         Image.ImageVisibility vis = Image.ImageVisibility.PUBLIC;
         Image im = Builders.imageV2()
                 .id(id)
